+++ superset-frontend/cypress-base/cypress/integration/chart_list/card_view.test.ts
@@ -92,11 +92,14 @@ describe('chart card view', () => {
     cy.get('[data-test="more-horiz"]').last().trigger('mouseover');
     cy.get('[data-test="chart-list-delete-option"]').should('be.visible');
     cy.get('[data-test="chart-list-delete-option"]').contains('Delete').click();
-    cy.get('[data-test="Please Confirm-modal-header"]').should('be.visible');
-    cy.get('[data-test="modal-delete-button"]').should('have.attr', 'disabled');
-    cy.get('[data-test="Please Confirm-modal-header"]').should('be.visible');
+    cy.get('[data-test="Please Confirm-modal"]').should('be.visible');
+    cy.get('[data-test="modal-confirm-button"]').should(
+      'have.attr',
+      'disabled',
+    );
+    cy.get('[data-test="Please Confirm-modal"]').should('be.visible');
     cy.get("[data-test='delete-modal-input']").type('DELETE');
-    cy.get('[data-test="modal-delete-button"]').should(
+    cy.get('[data-test="modal-confirm-button"]').should(
       'not.have.attr',
       'disabled',
     );
