+++ superset-frontend/cypress-base/cypress/integration/dashboard_list/card_view.test.ts
@@ -89,11 +89,11 @@ describe('Dashboard card view', () => {
     cy.get('.ant-dropdown-trigger').last().trigger('mouseover');
     cy.get('.ant-dropdown-menu-item').contains('Delete').should('exist');
     cy.get('.ant-dropdown-menu-item').contains('Delete').click();
-    cy.get('.modal-dialog').should('be.visible');
-    cy.get('.modal-dialog .btn-danger').should('have.attr', 'disabled');
-    cy.get(".modal-dialog input[id='delete']").type('DELETE');
-    cy.get('.modal-dialog .btn-danger').should('not.have.attr', 'disabled');
-    cy.get('.modal-dialog .btn-default').contains('Cancel').click();
+    cy.get('.ant-modal').should('be.visible');
+    cy.get('.ant-modal .btn-danger').should('have.attr', 'disabled');
+    cy.get(".ant-modal input[id='delete']").type('DELETE');
+    cy.get('.ant-modal .btn-danger').should('not.have.attr', 'disabled');
+    cy.get('.ant-modal .btn-default').contains('Cancel').click();
   });
 
   it('should edit correctly', () => {
