+++ superset-frontend/spec/javascripts/components/ConfirmStatusChange_spec.jsx
@@ -21,7 +21,7 @@ import { mount } from 'enzyme';
 import Button from 'src/components/Button';
 import { supersetTheme, ThemeProvider } from '@superset-ui/core';
 import ConfirmStatusChange from 'src/components/ConfirmStatusChange';
-import Modal from 'src/components/Modal';
+import Modal from 'src/common/components/Modal';
 
 describe('ConfirmStatusChange', () => {
   const mockedProps = {
