+++ superset-frontend/spec/javascripts/datasource/DatasourceModal_spec.jsx
@@ -18,7 +18,6 @@
  */
 import React from 'react';
 import { act } from 'react-dom/test-utils';
-import { Modal } from 'react-bootstrap';
 import configureStore from 'redux-mock-store';
 import { mount } from 'enzyme';
 import fetchMock from 'fetch-mock';
@@ -27,6 +26,7 @@ import sinon from 'sinon';
 import { supersetTheme, ThemeProvider } from '@superset-ui/core';
 
 import waitForComponentToPaint from 'spec/helpers/waitForComponentToPaint';
+import Modal from 'src/common/components/Modal';
 import DatasourceModal from 'src/datasource/DatasourceModal';
 import DatasourceEditor from 'src/datasource/DatasourceEditor';
 import * as featureFlags from 'src/featureFlags';
@@ -80,7 +80,7 @@ describe('DatasourceModal', () => {
   });
 
   it('renders a Modal', () => {
-    expect(wrapper.find(Modal)).toHaveLength(2);
+    expect(wrapper.find(Modal)).toExist();
   });
 
   it('renders a DatasourceEditor', () => {
