+++ superset-frontend/src/views/CRUD/data/dataset/AddDatasetModal.tsx
@@ -20,7 +20,7 @@ import React, { FunctionComponent, useState } from 'react';
 import { styled, SupersetClient, t } from '@superset-ui/core';
 import { isEmpty, isNil } from 'lodash';
 import Icon from 'src/components/Icon';
-import Modal from 'src/components/Modal';
+import Modal from 'src/common/components/Modal';
 import TableSelector from 'src/components/TableSelector';
 import withToasts from 'src/messageToasts/enhancers/withToasts';
 import { createErrorHandler } from 'src/views/CRUD/utils';
